Add airtable mvp
