+++ src/components/Nav.tsx
@@ -26,72 +26,12 @@ function Nav(props: Props) {
 						<div className="flex items-center">
 							<div className="flex-shrink-0">
 								<SecondaryButton
+									href="https://airtable.com/shrQgLHH0QV46ve7b"
+									openInNewWindow
 									leadingIcon={<Icon icon="plus-circle" className="w-6 mr-2" />}
 								>
 									List your place
 								</SecondaryButton>
-							</div>
-							<div className="hidden md:ml-4 md:flex-shrink-0 md:flex md:items-center">
-								<button className="bg-white p-1 rounded-full text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
-									<span className="sr-only">View notifications</span>
-									<Icon icon="bell" />
-								</button>
-								{/* Profile dropdown */}
-								<div className="ml-3 relative">
-									<div>
-										<button
-											className="bg-white rounded-full flex text-sm focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-											id="user-menu"
-											aria-haspopup="true"
-											onClick={() => setIsOpen(true)}
-										>
-											<span className="sr-only">Open user menu</span>
-											<img
-												className="h-8 w-8 rounded-full"
-												src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-											/>
-										</button>
-									</div>
-									<Transition
-										show={isOpen}
-										enter="transition ease-out duration-200"
-										enterFrom="transform opacity-0 scale-95"
-										enterTo="transform opacity-100 scale-100"
-										leave="transition ease-in duration-75"
-										leaveFrom="transform opacity-100 scale-100"
-										leaveTo="transform opacity-0 scale-95"
-									>
-										<div
-											className="origin-top-right absolute right-0 mt-2 w-48 rounded-md shadow-lg py-1 bg-white ring-1 ring-black ring-opacity-5"
-											role="menu"
-											aria-orientation="vertical"
-											aria-labelledby="user-menu"
-											ref={ref}
-										>
-											<a
-												href="#"
-												className="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-												role="menuitem"
-											>
-												Your Profile
-											</a>
-											<a
-												href="#"
-												className="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-												role="menuitem"
-											>
-												Settings
-											</a>
-											<a
-												href="#"
-												className="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-												role="menuitem"
-											>
-												Sign out
-											</a>
-										</div>
-									</Transition>
-								</div>
 							</div>
 						</div>
 					</div>
