+++ src/styles/tailwind.css
@@ -13,7 +13,6 @@
 body,
 html,
 #__next {
-	@apply dark:bg-black;
 }
 
 #__next {
